Fix crash when trying to add a new grid row while a first one was not yet saved. Closes #788

+++ source/main.pas
@@ -9401,8 +9401,12 @@ begin
       Allowed := Results.SaveModifications;
       DisplayRowCountStats(Sender);
     end else if Results.Inserted then begin
-      Results.DiscardModifications;
-      Sender.DeleteNode(OldNode);
+      if NewNode <> nil then begin
+        Results.DiscardModifications;
+        Sender.DeleteNode(OldNode);
+      end else begin
+        Allowed := False;
+      end;
     end;
   end;
 end;
